backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
requests==2.31.0
openai==1.30.1
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
python-multipart==0.0.9
email-validator==2.1.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
pymongo==4.6.3
motor==3.4.0
stripe==9.10.0
pandas==2.2.2
numpy==1.26.4
tiktoken==0.7.0
tenacity==8.2.3
rich==13.7.1
typer==0.12.3